Keep ruleset-specific `fix` sub-command tests in separate traits
